backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-dotenv==1.0.1
redis==5.2.1
